refactor: 트랜잭션 롤백 테스트 step4 요구사항에 맞게 변경
